Extract and push more methods down
